Move string utils from SparkSubmitOperator
