+++ tests/common/util/test_string_utils.py
@@ -1,0 +1,28 @@
+#
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+#
+import unittest
+
+from ai_flow.common.util.string_utils import mask_cmd
+
+
+class TestThreadUtils(unittest.TestCase):
+
+    def test_masks_passwords(self) -> None:
+        command_masked = mask_cmd(["spark-submit", "foo", "--bar", "baz", '--password=secret'])
+        self.assertEqual('spark-submit foo --bar baz --password=******', command_masked)
